added output, bug fixes, optimizations, improved makefile

+++ src/input_parser.cpp
@@ -10,6 +10,8 @@ input_parser::input_parser(const char* file_name)
 
 inline bool input_parser::string_to_uint32(const uint8_t* str,uint32_t len,uint32_t& value)
 {
+  //convert str to uint32
+  //return false if not an unsigned interger
   if(len == 0)
     return false;
   value = str[0]-'0';
@@ -17,7 +19,7 @@ inline bool input_parser::string_to_uint32(const uint8_t* str,uint32_t len,uint3
     return false;
   for(uint32_t i=1; i<len; ++i)
   {
-    uint8_t ch = str[i]-'0';
+    uint32_t ch = str[i]-'0';
     if(ch > 9)
       return false;
     value = 10*value+ch;
@@ -27,6 +29,7 @@ inline bool input_parser::string_to_uint32(const uint8_t* str,uint32_t len,uint3
 
 inline void input_parser::read_to_newline()
 {
+  //read until newline is reached
   while(true)
   {
     uint8_t ch = m_file.get();
